requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
tzdata
python-dotenv